init the About Us page
