+++ Frontend/src/views/About/organization.json
@@ -1,0 +1,12 @@
+[
+  {
+    "name": "Dr. Guy Yachdav", 
+    "role": "Supervisor & Initiator", 
+    "img": "https://scholar.googleusercontent.com/citations?view_op=medium_photo&user=UoUkGhUAAAAJ&citpid=2", 
+    "dscp": "Technology executive with over 15 years experience in R&D and specialization in big data and machine learning", 
+    "socialFB": "", 
+    "socialLinkedIn": "https://www.linkedin.com/in/gyachdav/?originalSubdomain=il", 
+    "socialGithub": "", 
+    "socialTwitter": ""
+  }
+]
